requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
postgres==3.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pytz==2021.1
sqlparse==0.4.1
